+++ src/main/java/com/wc/watchu/repository/MovieRepository.java
@@ -8,9 +8,9 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MovieRepository extends JpaRepository<MovieEntity, Long> {
-    boolean existsByContent_ContentsId(Long contentsId);
-    List<MovieEntity> findByContent_ContentsId(Long contentsId);
+public interface MovieRepository extends JpaRepository<MovieEntity, Integer> {
+    boolean existsByContent_ContentsId(int contentsId);
+    List<MovieEntity> findByContent_ContentsId(int contentsId);
 
     @Query("SELECT m FROM MovieEntity m ORDER BY m.views DESC")
     List<MovieEntity> findTop5ByViewsDesc();
